페이징 기능 튜닝 - @Query 만들어 성능 최적화

+++ src/main/java/kr/spartaclub/scheduleproject/dto/schedule/GetPageableScheduleResponse.java
@@ -9,12 +9,12 @@ public class GetPageableScheduleResponse {
     private final Long id;
     private final String title;
     private final String content;
-    private final int commentCount;
+    private final Long commentCount;
     private final String name;
     private final LocalDateTime createdAt;
     private final LocalDateTime modifiedAt;
 
-    public GetPageableScheduleResponse(Long id, String title, String content, int commentCount, String name, LocalDateTime createdAt, LocalDateTime modifiedAt) {
+    public GetPageableScheduleResponse(Long id, String title, String content, Long commentCount, String name, LocalDateTime createdAt, LocalDateTime modifiedAt) {
         this.id = id;
         this.title = title;
         this.content = content;
